% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_peak_finding.R
\name{find_peaks}
\alias{find_peaks}
\title{Find the location of clusters of points from a set of 2D points}
\usage{
find_peaks(opts, prop_cutoff = 0.7)
}
\arguments{
\item{opts}{An options object created using make_options()}

\item{prop_cutoff}{The lowest relative density, 0<cut<1 which is considered to be a peak. Should be a multiple of 0.1}
}
\value{
A list of TPeaks. Each TPeak is a list of clusters at some threshold T. Each cluster has \verb{$ij} and \verb{$xy} attributes, which are the index and position information of all points that are othogonally connected to form a true peak at threshold T.
}
\description{
A continuous density distribution is generated by convolving the spread of points
with a Gaussian kernel. This distribution is sampled on a grid (like pixels),
and the pixel heights are normalized to have a min/max of 0 and 1. The heights
are discretized to multiples of 0.1, and at each of these thresholds a flood-fill
algorithm finds the location of pixels corresponding to a peak at that threshold.
Orthogonally-connected pixels are clustered together.
}
