% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_options.R
\name{make_opts}
\alias{make_opts}
\title{Creates an options object with information about the grid, given various parameters}
\usage{
make_opts(
  x_positions,
  y_positions,
  resolution = 0.1,
  smear_size = -1,
  x_extent = c(0, 0),
  y_extent = c(0, 0)
)
}
\arguments{
\item{x_positions}{A list of x-coordinates for all points}

\item{y_positions}{A list of y-coordinates for all points}

\item{resolution}{The width of a pixel}

\item{smear_size}{The standard deviation of the Gaussian kernel that convolves with the datapoints to create a 2D density distribution}

\item{x_extent}{A length-2 vector of the minimum and maximum x-values of the grid}

\item{y_extent}{A length-2 vector of the minimum and maximum y-values of the grid}
}
\value{
A list of various properties of the grid. In order, xMin, xMax, yMin, yMax, nX, nY, resolution, smear_size, x_positions, y_positions, n_points, grid. The last, \code{grid}, is an nY x nX array filled with zeroes.
}
\description{
Collects together all the relevant parameters for peak-finding and visualization. These are
bundled together into an \code{options} object so these parameters don't need to be passed
around individually
}
