% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_plotting.R
\name{plot_peaks}
\alias{plot_peaks}
\title{Plot all peaks as ellipsoids}
\usage{
plot_peaks(
  peaks,
  opts,
  new_plot = TRUE,
  overtitle = "Peak Locations",
  show_frame = FALSE
)
}
\arguments{
\item{peaks}{The object that was returned from find_peaks()}

\item{opts}{The options object that was passed to find_peaks()}

\item{new_plot}{Boolean for whether a new plot should be created. Use FALSE to continue writing to the current plot.}

\item{overtitle}{The main title of the new plot}

\item{show_frame}{Boolean of whether to show the bounding region and xy axes}
}
\description{
Using the \code{peaks} object that is returned from the \code{find_peaks()} function,
plot each cluster as an ellipsoid based off the constituent pixels of that cluster.
The original datapoints are overlaid as an inefficient scatterplot.
}
