% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_peak_finding.R
\name{print_peaks}
\alias{print_peaks}
\title{Prints the results of find_peaks() in a nice way}
\usage{
print_peaks(peaks, opts)
}
\arguments{
\item{peaks}{The output of find_peaks()}

\item{opts}{The output of make_opts()}
}
\description{
Prints the results of find_peaks() in a nice way
}
