% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbd_plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap of 2D point dataset}
\usage{
plot_heatmap(opts, overtitle = "Point Heatmap", show_frame = FALSE)
}
\arguments{
\item{opts}{An options object created from make_grid()}

\item{overtitle}{The main title of the new plot}

\item{show_frame}{Boolean for whether to show the bounding region and xy axes}
}
\description{
Plot heatmap of 2D point dataset
}
